requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.6
langchain-openai==0.1.13
langchain-community==0.2.6
langgraph==0.1.9
openai==1.35.3
twilio==9.2.3
apscheduler==3.10.4
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
alembic==1.13.1
pydantic==2.7.4
pydantic-settings==2.3.3
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.134.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
pytz==2024.1
python-dateutil==2.9.0.post0
